Add macro support for register arrays
I'm experimenting with a RAL code generator that collapses contiguous
register arrays. The generated code would resemble

    pub struct RegisterBlock {
        pub MY_ARRAY: [RWRegister<u32>; 3],
    }

and an individual register would be addressed like

    ral::read_reg!(ral::my_mod, my_inst, MY_ARRAY[1]);

This commit extends the four macros so that we can specify an array
offset. We simply need to match zero or more `[N]` patterns, where `N`
is some expression that produces an array offset. The included test case
shows that the approach should support multi-dimensional arrays.