Document support for register arrays in each macro
Since stm32ral doesn't have register arrays, I'm using prose and
pseudo-code to document each macro's support for register arrays. This
seemed to be today's simplest way to document the features for users.

I have another commit that implements a hidden module with a RAL-like
API and register arrays. Documentation examples then reference this
hidden module to demonstrate the register array feature. But, it adds
more code (could be shared with the tests), and it resulted in
inconsistent documentation examples when compared to the STM examples.